Winrm_health_check/requirements.txt
cloudshell-shell-core>=5.0.3,<6.0.0
pywinrm>0.4.0,<0.5.0